Moving LinearLookupFunction to its own class
