+++ src/main/kotlin/net/maizegenetics/phgv2/pathing/ropebwt/SplineUtils.kt
@@ -14,7 +14,6 @@ import org.apache.commons.math3.analysis.interpolation.AkimaSplineInterpolator
 import org.apache.commons.math3.analysis.polynomials.PolynomialSplineFunction
 import org.apache.logging.log4j.LogManager
 import java.io.*
-import kotlin.math.abs
 import kotlin.random.Random
 
 @Serializable
@@ -748,53 +747,6 @@ class SplineUtils{
             knotMap: RangeMap<Position,Pair<Position,Position>>
         ) : LinearLookupFunction {
             return LinearLookupFunction(knotMap)
-        }
-    }
-}
-
-
-class LinearLookupFunction(val knotMap: RangeMap<Position,Pair<Position,Position>>) {
-    fun value(position: Position): Position {
-        //Find the range that contains the position
-
-        val range = knotMap.getEntry(position)
-        return if(range != null) {
-            val asmRange = range.key
-            val asmSt = asmRange.lowerEndpoint()
-            val asmEnd = asmRange.upperEndpoint()
-            resolveLinearInterpolation(position,Pair(asmSt,asmEnd), range.value)
-        } else {
-            Position("unknown", 0) // Return a default position if not found
-        }
-    }
-
-    fun resolveLinearInterpolation(
-        asmPosition: Position,
-        asmRange : Pair<Position, Position>,
-        referenceRange : Pair<Position, Position>
-    ): Position {
-
-        val offsetProp = (asmPosition.position - asmRange.first.position).toDouble() /
-                (asmRange.second.position - asmRange.first.position).toDouble()
-
-        val offsetPos = abs(referenceRange.second.position - referenceRange.first.position) * offsetProp
-
-        if(referenceRange.first.contig != referenceRange.second.contig) {
-            //If the contigs are not the same, we cannot interpolate
-            Position("unknown", 0)
-        }
-        return if(referenceRange.first.position == referenceRange.second.position) {
-            //If the positions are the same, all values are equal
-            Position(referenceRange.first.contig, referenceRange.first.position)
-        }
-        else if(referenceRange.first.position < referenceRange.second.position) {
-            //Positive strand
-            val newPos = referenceRange.first.position + offsetPos
-            Position(referenceRange.first.contig, newPos.toInt())
-        } else {
-            //Negative strand
-            val newPos = referenceRange.first.position - offsetPos
-            Position(referenceRange.first.contig, newPos.toInt())
         }
     }
 }
